requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
coverage==7.4.0
Django==5.0.1
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==22.5.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
psycopg2==2.9.9
PyJWT==2.8.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
